+++ azure/datalake/store/__init__.py
@@ -6,7 +6,8 @@
 # license information.
 # --------------------------------------------------------------------------
 
-__version__ = "0.0.32"
+__version__ = "0.0.34"
+
 
 from .core import AzureDLFileSystem
 from .multithread import ADLDownloader