+++ azure/datalake/store/core.py
@@ -19,6 +19,7 @@ import io
 import logging
 import sys
 import uuid
+import json
 
 
 # local imports
@@ -552,11 +553,14 @@ class AzureDLFileSystem(object):
             directory, and delete that whole directory when done.
         """
         outfile = AzureDLPath(outfile).trim()
-        filelist = ','.join(AzureDLPath(f).as_posix() for f in filelist)
         delete = 'true' if delete_source else 'false'
+        sourceList = [AzureDLPath(f).as_posix() for f in filelist]
+        sources = {}
+        sources["sources"] = sourceList
         self.azure.call('MSCONCAT', outfile.as_posix(),
-                        data='sources='+filelist,
-                        deleteSourceDirectory=delete)
+                        data=bytearray(json.dumps(sources,separators=(',', ':')), encoding="utf-8"),
+                        deleteSourceDirectory=delete,
+                        headers={'Content-Type': "application/json"},)
         self.invalidate_cache(outfile)
 
     merge = concat
@@ -869,6 +873,20 @@ class AzureDLFile(object):
         return out
 
     read1 = read
+
+    def readinto(self, b):
+        """
+        Reads data into buffer b
+        Returns number of bytes read.
+
+        Parameters
+        ----------
+        b : bytearray
+            Buffer to which bytes are read into
+        """
+        temp = self.read(len(b))
+        b[:len(temp)] = temp
+        return len(temp)
 
     def write(self, data):
         """